+++ src/main/java/com/green/muziuniv_be_notuser/app/shared/course/model/CourseFilterRes.java
@@ -17,7 +17,7 @@ public class CourseFilterRes {
     private int year; //연도
     private int semester; // 학기
     private String time; // 강의시간
-    private int credit; // 학점
+    private Long credit; // 학점
     private int maxStd; // 정원
     private int remStd; //잔여 인원
     private String professorName; // 담당 교수 이름
