+++ src/main/java/com/green/muziuniv_be_notuser/app/student/enrollment/model/EnrollmentRes.java
@@ -22,7 +22,7 @@ public class EnrollmentRes {
     private int year; //연도
     private int semester; // 학기
     private String time; // 강의시간
-    private int credit; // 학점
+    private Long credit; // 학점
     private int maxStd; // 정원
     private int remStd; //잔여 인원
     private String deptName; // 개설학과
